рефакторинг

+++ paymentAPI/src/main/kotlin/com/raif/paymentapi/domain/dto/SubscriptionPaymentDto.kt
@@ -1,8 +1,10 @@
 package com.raif.paymentapi.domain.dto
 
+import java.math.BigDecimal
+
 data class SubscriptionPaymentDto(
     val additionalInfo: String?,
-    val amount: Long,
+    val amount: BigDecimal,
     val currency: String?,
     val order: String
 )
